requirements.txt
Django==2.0.5
pytz==2018.4
celery
redis
psycopg2
djangorestframework
dj-database-url
graypy
